+++ src/mlp.h
@@ -15,8 +15,8 @@ typedef enum {
 }layer_type;
 
 typedef struct {
-    void (*f)(double* inputs, double* outputs, int len);
-    void (*df)(double* inputs, double* outputs, int len);
+    void (*f)(float* inputs, float* outputs, int len);
+    void (*df)(float* inputs, float* outputs, int len);
     fn_name function_name;
 }function;
 
@@ -30,12 +30,12 @@ typedef struct {
     layer_type type;
     int n_inputs;           // = Input shape = Number of weights
     int n_outputs;          // Number of neurons
-    double* weights;        // List of all the weights of all the neurons in the layer
-    double* biases;         // Bias of each neuron in the layer
-    double* raw_outputs;    // Outputs of all the neurons
-    double* outputs;        // Outputs of all the neurons after the activation function
-    double* derivatives;    // Outputs of all the neurons after the activation function's derivative (for training)
-    double* deltas;         // Error of each neuron (for training)
+    float* weights;        // List of all the weights of all the neurons in the layer
+    float* biases;         // Bias of each neuron in the layer
+    float* raw_outputs;    // Outputs of all the neurons
+    float* outputs;        // Outputs of all the neurons after the activation function
+    float* derivatives;    // Outputs of all the neurons after the activation function's derivative (for training)
+    float* deltas;         // Error of each neuron (for training)
     function* activation_function;
 }layer;
 
@@ -45,15 +45,15 @@ typedef struct {
 } MLP;
 
 
-double ranged_rand(double min, double max);
+float ranged_rand(float min, float max);
 layer dense(int n_neurons, int n_inputs, function *activation_function);
-void forward(MLP *m, double* inputs, int n_inputs);
-void train(MLP *m, double* raw_inputs, double* target, double lr);
+void forward(MLP *m, float* inputs, int n_inputs);
+void train(MLP *m, float* raw_inputs, float* target, float lr);
 int get_num_parameters(MLP* mlp);
 void print_model(MLP* m);
-void print_output(MLP *m, double* input, int input_len, double *expected, int expected_len);
-void print_list(double* list, int len);
+void print_output(MLP *m, float* input, int input_len, float *expected, int expected_len);
+void print_list(float* list, int len);
 void free_model(MLP* m);
-void one_hot(double* output, int input, int n_classes);
+void one_hot(float* output, int input, int n_classes);
 void save_model(MLP* m, const char* path);
 void load_model(MLP* m, const char* path);
